pyproject.toml
[project]
name = "clustera-api-kernel"
version = "0.1.0"
description = "Clustera API kernel - REST API endpoint for submitting natural language queries to Kafka"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "kafka-python",
    "python-dotenv",
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
]

[project.optional-dependencies]
# Optional dependencies for enhanced functionality
monitoring = [
    "prometheus-client",     # Metrics collection
]
security = [
    "python-jose[cryptography]",  # JWT handling
    "passlib[bcrypt]",       # Password hashing
]
database = [
    "psycopg2-binary",      # PostgreSQL (if needed for logging/state)
    "sqlalchemy",           # ORM
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]